+++ lib/kitchen/directions/bake_injected_exercise/add_injected_exercise_id.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Kitchen
+  module Directions
+    # Ids should be added before exercises are moved to EOC,
+    # since they're using part of the ancestor page id.
+    #
+    # In some books exercises are numbered after moving.
+    # That's why this step has to be separated from BakeInjectedExerciseQuestion
+    module AddInjectedExerciseId
+      def self.v1(question:)
+        question.id
+      end
+    end
+  end
+end