+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Modify `BakeAutotitledNotes` to bake unnumbered exercises with solution (minor)
+* Create `AddInjectedExerciseId` to separate creating ids from `BakeInjectedExerciseQuestion` (minor)
+
 ## [12.1.0] - 2021-09-24
 
 * Fix `BakeExample#titles_to_rename` to exclude exercise titles (patch)